refactor: move _write_survey_participants into survey_participants_logic.py
